Add all-to-all processors
This commit enables creating all-to-all processors, i.e. processors
without coupling map. This will be necessary for the logical models,
which enable all-to-all connectivity between the logical qubits.

Why is this necessary? Basically because it was impossible to combine
the Qiskit layout transpilation passes (required when a coupling map
exists) and the synthesis algorithms (required when the gate set
contains only discrete gates, which is the case for logical qubits).
As a side note, this really shows how Qiskit is a messy codebase.

This commit creates an unwelcome complexity in the code (because the
cases of instructions tied to qubits and of instructions available on
all qubits must be handled in the same code), but I believe this is
still manageable.