qir_to_qiskit: change name of instructions
The name of instructions should not be an arbitrary string reflecting
the true nature of the instruction, but rather an existing Qiskit
intstruction.
We knew that already in local/, but since the next commits will make
remote/ perform actual transpilation before sending the circuit to the
API, this problem became apparent.