remote: use SKSynthesisPlugin when possible
The remote/ backend now uses the SKSynthesisPlugin when (1)
rotations are absent from the gate set and (2) H and T are in the
gate set.
This will allow us to support logical qubit targets through the API.
The drawback is that the decision rule (1 + 2 above) may not be suitable
for some given situation in the future. Time will tell!