local: fix ProcessorSimulator for large processors
Simulation used to fail when the number of qubits of the processor was
larger than the number of qubits supported by AerSimulator.
This commit fixes this.

Note that this is caused by the release qiskit-aer==0.13.0, and we
forbid this version in the next commit because it is too slow.
Nevertheless, the fix is valuable in itself, for the day 0.13.1 comes
out or if we find a fix for the slowness.