remote: send native operations to the API
This commit, by using a new version of qiskit-qir-alice-bob-fork,
removes a significant piece of technical debt:
* The version of qiskit-qir now supports native operations like mx,
  prepare_x, prepare_z
* This makes it possible for the remote backend to transpile to native
  A&B qir and send it directly to the API.

Additionally, this commit factorizes the LocalStatePreparationPlugin and
the RemoteStatePreparationPlugin into a single StatePreparationPlugin.
As a consequence, this means that remote backends now cause Qiskit
to perform some transpilation when a circuit is executed.
This will allow us to support through the API a target like
EMU:6Q:PHYSICAL_CATS.