remote: Allow passing of options in get_backend()
When we introduced the local provider, we made it so we declare
its options (nb_photons, kappa_1...) when using get_backend().

This is contradictory to what is happening in the remote provider,
where options must be passed in execute().

This commit harmonizes this behaviour by allowing to set the options
in get_backend() for the remote provider.
(which actually makes some sense to do it like this when we want to
set the nb_photons for example).
This allows to quickly swap between the local and remote provider.

We aim to stay as permissive as possible, by still letting the user
passing options at execute() for the remote provider.