remote: Use Pydantic's camel & snake case functions
This is a fix where we found out the "kappa1" parameter of targets
was not converted to "kappa_1" in snake_case when getting a backend as
it should.

This came from our home made to_snake_case and to_camel_case functions
using some obscure regex. Instead, this commit replaces these functions
with Pydantic's, that are more tested and proven.