fix: Allow transpilation of circuits multiple qregs.
We were not properly mapping the virtual to physical qubits in our
custom pass manager, which resulted in the impossibility of using
the full capabilities of our backends.

Indeed, this prevented from defining circuits with multiple quantum
registries that Qiskit would have been able to a single physical
registry to run on.

This commit addresses that, adding a Layout when applying the custom
pass manager of our backends to ensure we can map qubits to physical
registries.