fix(sk_synthesis): Remove cz & ccz from synth gates
We used to apply the SK synthesis algorithm to the CZ & CCZ gates.
This behaviour was inappropriate as these gates are discrete gates,
and should not be approximated like a continuous one (e.g.: rZ(theta)).

The CZ & CCZ should rather be replaced by CX & CCX respectively in the
"Basis Translation" step of transpilation instead.

This commit does this by removing CZ & CCZ from the list of gates to
synthesize, so they will be ignored by the SK synthesis step.