fix: measure_x custom instruction transpiling
The function substitute_node_with_dag is patched to make the custom
measure_x instruction work with Qiskit 1.2.
It adds the else statement that was removed on lines 1610-1611 in
dagcircuit.py in this commit :
https://github.com/Qiskit/qiskit/commit/353b0ea6bdd907e801ad8fa264f3444e0be942aa#diff-4fb31a3ade5ae57cfd91ea00dbf3c5b6ab066a8234a742d91f9c09a09edca2f7L1610-L1611
This function will be moved to Rust in Qiskit version 1.3 so we will
have to find another solution.