fix(remote): backend default options not set properly in remote provider
When using AliceBobRemoteProvider.get_backend() with parameters, a
Python reference of the backend was used and the parameters were
modified on the object of the AliceBobRemoteProvider instance created
earlier. It was causing the default options of a backend to be
overwritten definitely for an AliceBobRemoteProvider instance when
calling get_backend() with parameters.

These changes use a new instance of the backend object instead of a
reference to the same instance to make each backend returned by
get_backend() and backends() totally independent (as get_backends()
uses backends() under the hood). We need to recreate the backend
instance each time and copy.deepcopy() was not used to avoid any
potential issues with cyclic graph or things like that.