fix(remote): shared reference of options in backend.run()
A new instance of Options is needed to allow to modify the options
temporarily. We need to call Options.update_options() without definitely
modifying the options of the backend through a shared reference.

Some code is factorized under update_options_object() to allow to update
an Options object different from the AliceBobRemoteBackend.options
attribute.