feat: update to qiskit 1.3
Some tests had to be updated for Qiskit 1.3.
This new Qiskit version changes the default optimization_level of
transpile() from 1 to 2. This change enables other
transpilation passes like VF2Layout and ApplyLayout passes that shuffle
the qbits randomly in the circuit. As we need a deterministic and
reproducible behavior for the tests, we fix the seed in transpile().