fix: issue when options are passed as numpy types
When using the AliceBobRemoteProvider, the backend options are serialized as JSON to be sent to the API.
In some cases, the user might provide some option values as numpy types (eg. `average_nb_photons=np.int32(4)`), which is not JSON serializable and therefore raises an error.

This fix aims to convert numpy scalar values to their corresponding native Python type before serialization.
`value.item()` does just that.